+++ API/Collection/UserCollection.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Kaliop\eZMigrationBundle\API\Collection;
+
+/**
+ * @todo add phpdoc to suggest typehinting
+ */
+class UserCollection extends AbstractCollection
+{
+    protected $allowedClass = 'eZ\Publish\API\Repository\Values\User\User';
+}