+++ Core/Executor/UserManager.php
@@ -2,12 +2,22 @@
 
 namespace Kaliop\eZMigrationBundle\Core\Executor;
 
+use Kaliop\eZMigrationBundle\API\Collection\UserCollection;
+use Kaliop\eZMigrationBundle\Core\Matcher\UserMatcher;
+
 /**
  * Handles user migrations.
  */
 class UserManager extends RepositoryExecutor
 {
     protected $supportedStepTypes = array('user');
+
+    protected $userMatcher;
+
+    public function __construct(UserMatcher $userMatcher)
+    {
+        $this->userMatcher = $userMatcher;
+    }
 
     /**
      * Creates a user based on the DSL instructions.
@@ -59,6 +69,8 @@ class UserManager extends RepositoryExecutor
         $user = $userService->createUser($userCreateStruct, $userGroups);
 
         $this->setReferences($user);
+
+        return $user;
     }
 
     /**
@@ -68,65 +80,71 @@ class UserManager extends RepositoryExecutor
      */
     protected function update()
     {
-        if (!isset($this->dsl['id'])) {
-            throw new \Exception('No user set for update.');
+        $userCollection = $this->matchUsers('user');
+
+        if (count($userCollection) > 1 && array_key_exists('references', $this->dsl)) {
+            throw new \Exception("Can not execute User update because multiple user match, and a references section is specified in the dsl. References can be set when only 1 user matches");
+        }
+
+        if (count($userCollection) > 1 && array_key_exists('email', $this->dsl)) {
+            throw new \Exception("Can not execute User update because multiple user match, and an email section is specified in the dsl.");
         }
 
         $userService = $this->repository->getUserService();
 
-        $userId = $this->dsl['id'];
-        if ($this->referenceResolver->isReference($userId)) {
-            $userId = $this->referenceResolver->getReferenceValue($userId);
-        }
+        foreach($userCollection as $key => $user) {
 
-        $user = $userService->loadUser($userId);
+            $userUpdateStruct = $userService->newUserUpdateStruct();
 
-        $userUpdateStruct = $userService->newUserUpdateStruct();
-
-        if (isset($this->dsl['email'])) {
-            $userUpdateStruct->email = $this->dsl['email'];
-        }
-        if (isset($this->dsl['password'])) {
-            $userUpdateStruct->password = (string)$this->dsl['password'];
-        }
-        if (isset($this->dsl['enabled'])) {
-            $userUpdateStruct->enabled = $this->dsl['enabled'];
-        }
-
-        $userService->updateUser($user, $userUpdateStruct);
-
-        if (isset($this->dsl['groups'])) {
-            if (!is_array($this->dsl['groups'])) {
-                $this->dsl['groups'] = array($this->dsl['groups']);
+            if (isset($this->dsl['email'])) {
+                $userUpdateStruct->email = $this->dsl['email'];
+            }
+            if (isset($this->dsl['password'])) {
+                $userUpdateStruct->password = (string)$this->dsl['password'];
+            }
+            if (isset($this->dsl['enabled'])) {
+                $userUpdateStruct->enabled = $this->dsl['enabled'];
             }
 
-            $assignedGroups = $userService->loadUserGroupsOfUser($user);
+            $user = $userService->updateUser($user, $userUpdateStruct);
 
-            // Assigning new groups to the user
-            foreach ($this->dsl['groups'] as $groupToAssignId) {
-                $present = false;
-                foreach ($assignedGroups as $assignedGroup) {
-                    // Make sure we assign the user only to groups he isn't already assigned to
-                    if ($assignedGroup->id == $groupToAssignId) {
-                        $present = true;
-                        break;
+            if (isset($this->dsl['groups'])) {
+                if (!is_array($this->dsl['groups'])) {
+                    $this->dsl['groups'] = array($this->dsl['groups']);
+                }
+
+                $assignedGroups = $userService->loadUserGroupsOfUser($user);
+
+                // Assigning new groups to the user
+                foreach ($this->dsl['groups'] as $groupToAssignId) {
+                    $present = false;
+                    foreach ($assignedGroups as $assignedGroup) {
+                        // Make sure we assign the user only to groups he isn't already assigned to
+                        if ($assignedGroup->id == $groupToAssignId) {
+                            $present = true;
+                            break;
+                        }
+                    }
+                    if (!$present) {
+                        $groupToAssign = $userService->loadUserGroup($groupToAssignId);
+                        $userService->assignUserToUserGroup($user, $groupToAssign);
                     }
                 }
-                if (!$present) {
-                    $groupToAssign = $userService->loadUserGroup($groupToAssignId);
-                    $userService->assignUserToUserGroup($user, $groupToAssign);
+
+                // Unassigning groups that are not in the list in the migration
+                foreach ($assignedGroups as $assignedGroup) {
+                    if (!in_array($assignedGroup->id, $this->dsl['groups'])) {
+                        $userService->unAssignUserFromUserGroup($user, $assignedGroup);
+                    }
                 }
             }
 
-            // Unassigning groups that are not in the list in the migration
-            foreach ($assignedGroups as $assignedGroup) {
-                if (!in_array($assignedGroup->id, $this->dsl['groups'])) {
-                    $userService->unAssignUserFromUserGroup($user, $assignedGroup);
-                }
-            }
+            $userCollection[$key] = $user;
         }
 
-        $this->setReferences($user);
+        $this->setReferences($userCollection);
+
+        return $userCollection;
     }
 
     /**
@@ -134,48 +152,64 @@ class UserManager extends RepositoryExecutor
      */
     protected function delete()
     {
-        if (!isset($this->dsl['user_id']) && !isset($this->dsl['email']) && !isset($this->dsl['username'])) {
-            throw new \Exception('No user set for update.');
-        }
+        $userCollection = $this->matchUsers('delete');
 
         $userService = $this->repository->getUserService();
 
-        $usersToDelete = array();
-
-        /// @todo add support for alias 'id'
-        if (isset($this->dsl['user_id'])) {
-            if (!is_array($this->dsl['user_id'])) {
-                $this->dsl['user_id'] = array($this->dsl['user_id']);
-            }
-            foreach ($this->dsl['user_id'] as $userId) {
-                if ($this->referenceResolver->isReference($userId)) {
-                    $userId = $this->referenceResolver->getReferenceValue($userId);
-                }
-                array_push($usersToDelete,$userService->loadUser($userId));
-            }
-        }
-
-        if (isset($this->dsl['email'])) {
-            if (!is_array($this->dsl['email'])) {
-                $this->dsl['email'] = array($this->dsl['email']);
-            }
-            foreach ($this->dsl['email'] as $email) {
-                $usersToDelete = array_merge($usersToDelete,$userService->loadUsersByEmail($email));
-            }
-        }
-
-        if (isset($this->dsl['username'])) {
-            if (!is_array($this->dsl['username'])) {
-                $this->dsl['username'] = array($this->dsl['username']);
-            }
-            foreach ($this->dsl['username'] as $username) {
-                array_push($usersToDelete,$userService->loadUserByLogin($username));
-            }
-        }
-
-        foreach($usersToDelete as $user) {
+        foreach($userCollection as $user) {
             $userService->deleteUser($user);
         }
+
+        return $userCollection;
+    }
+
+    /**
+     * @param string $action
+     * @return RoleCollection
+     * @throws \Exception
+     */
+    protected function matchUsers($action)
+    {
+        if (!isset($this->dsl['id']) && !isset($this->dsl['user_id']) && !isset($this->dsl['email']) && !isset($this->dsl['username']) && !isset($this->dsl['match'])) {
+            throw new \Exception("The id, email or username of a user or a match condition is required to $action it.");
+        }
+
+        // Backwards compat
+        if (!isset($this->dsl['match'])) {
+            $conds = array();
+            if (isset($this->dsl['id'])) {
+                 $conds['id'] = $this->dsl['id'];
+            }
+            if (isset($this->dsl['user_id'])) {
+                $conds['id'] = $this->dsl['user_id'];
+            }
+            if (isset($this->dsl['email'])) {
+                $conds['email'] = $this->dsl['email'];
+            }
+            if (isset($this->dsl['username'])) {
+                $conds['login'] = $this->dsl['username'];
+            }
+            $this->dsl['match'] = $conds;
+        }
+
+        $match = $this->dsl['match'];
+
+        // convert the references passed in the match
+        foreach ($match as $condition => $values) {
+            if (is_array($values)) {
+                foreach ($values as $position => $value) {
+                    if ($this->referenceResolver->isReference($value)) {
+                        $match[$condition][$position] = $this->referenceResolver->getReferenceValue($value);
+                    }
+                }
+            } else {
+                if ($this->referenceResolver->isReference($values)) {
+                    $match[$condition] = $this->referenceResolver->getReferenceValue($values);
+                }
+            }
+        }
+
+        return $this->userMatcher->match($match);
     }
 
     /**
@@ -183,7 +217,7 @@ class UserManager extends RepositoryExecutor
      *
      * The User manager currently only supports setting references to user_id.
      *
-     * @param \eZ\Publish\API\Repository\Values\User\User $user
+     * @param \eZ\Publish\API\Repository\Values\User\User|UserCollection $user
      * @throws \InvalidArgumentException when trying to set references to unsupported attributes
      * @return boolean
      */
@@ -191,6 +225,13 @@ class UserManager extends RepositoryExecutor
     {
         if (!array_key_exists('references', $this->dsl)) {
             return false;
+        }
+
+        if ($user instanceof UserCollection) {
+            if (count($user) > 1) {
+                throw new \InvalidArgumentException('User Manager does not support setting references for creating/updating of multiple users');
+            }
+            $user = reset($user);
         }
 
         foreach ($this->dsl['references'] as $reference) {